Secure Test Removal Updates/Changes
- Removed Secure Tests(900+)
- Removed test_m002(606)
- Moved test_c037 -> test_c018
- Moved test_c038 -> test_c028
