--- platform/secure_sw/arm-tf/aarch64/sbsa_helpers.h
@@ -1,27 +1,0 @@
-/** @file
- * Copyright (c) 2016, ARM Limited or its affiliates. All rights reserved.
-
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-
-#ifndef __SBSA_HELPERS_H__
-#define __SBSA_HELPERS_H__
-
-#include "arch_helpers.h"
-
-/**
-  @brief  System register access prototypes
- **/
-DEFINE_SYSREG_RW_FUNCS(mdcr_el3)
-
-#endif /* __SBSA_HELPERS_H__ */
