+++ uefi_app/SbsaAvsMain.c
@@ -1,5 +1,5 @@
 /** @file
- * Copyright (c) 2016-2020, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2016-2021, Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
 
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -33,7 +33,6 @@
 UINT32  g_sbsa_level;
 UINT32  g_enable_pcie_tests;
 UINT32  g_print_level;
-UINT32  g_execute_secure;
 UINT32  g_execute_nist;
 UINT32  g_skip_test_num[3] = {10000, 10000, 10000};
 UINT32  g_sbsa_tests_total;
@@ -250,14 +249,13 @@ HelpMsg (
   VOID
   )
 {
-  Print (L"\nUsage: Sbsa.efi [-v <n>] | [-l <n>] | [-f <filename>] | [-s] | [-skip <n>] | [-nist] | [-p <n>]\n"
+  Print (L"\nUsage: Sbsa.efi [-v <n>] | [-l <n>] | [-f <filename>] | [-skip <n>] | [-nist] | [-p <n>]\n"
          "Options:\n"
          "-v      Verbosity of the Prints\n"
          "        1 shows all prints, 5 shows Errors\n"
          "-l      Level of compliance to be tested for\n"
          "        As per SBSA spec, 0 to 6\n"
          "-f      Name of the log file to record the test results in\n"
-         "-s      Enable the execution of secure tests\n"
          "-skip   Test(s) to be skipped\n"
          "        Refer to section 4 of SBSA_ACS_User_Guide\n"
          "        To skip a module, use Model_ID as mentioned in user guide\n"
@@ -272,7 +270,6 @@ STATIC CONST SHELL_PARAM_ITEM ParamList[] = {
   {L"-v"    , TypeValue},    // -v    # Verbosity of the Prints. 1 shows all prints, 5 shows Errors
   {L"-l"    , TypeValue},    // -l    # Level of compliance to be tested for.
   {L"-f"    , TypeValue},    // -f    # Name of the log file to record the test results in.
-  {L"-s"    , TypeFlag},     // -s    # Binary Flag to enable the execution of secure tests.
   {L"-skip" , TypeValue},    // -skip # test(s) to skip execution
   {L"-help" , TypeFlag},     // -help # help : info about commands
   {L"-h"    , TypeFlag},     // -h    # help : info about commands
@@ -326,13 +323,6 @@ ShellAppMainsbsa (
               g_skip_test_num[++j] = StrDecimalToUintn((CONST CHAR16 *)(CmdLineArg+i+1));
           }
       }
-  }
-
-  // Options with Flag
-  if (ShellCommandLineGetFlag (ParamPackage, L"-s")) {
-    g_execute_secure = TRUE;
-  } else {
-    g_execute_secure = FALSE;
   }
 
   // Options with Values
@@ -434,12 +424,6 @@ ShellAppMainsbsa (
   val_pe_context_save(AA64ReadSp(), (uint64_t)branch_label);
   val_pe_initialize_default_exception_handler(val_pe_default_esr);
   FlushImage();
-
-  if (g_execute_secure == TRUE) {
-    Print(L"\n      ***  Starting Secure FW tests ***  \n");
-    val_secure_execute_tests(g_sbsa_level, val_pe_get_num());
-    Print(L"\n      ***  Secure FW tests completed ***  \n");
-  }
 
   Print(L"\n      ***  Starting PE tests ***  \n");
   Status = val_pe_execute_tests(g_sbsa_level, val_pe_get_num());
