+++ val/include/sbsa_avs_peripherals.h
@@ -1,5 +1,5 @@
 /** @file
- * Copyright (c) 2016-2018, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2016-2021, Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
 
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -40,6 +40,5 @@ uint32_t d003_entry(uint32_t num_pe);
 
 
 uint32_t m001_entry(uint32_t num_pe);
-uint32_t m002_entry(uint32_t num_pe);
 
 #endif // __SBSA_AVS_PERIPHERAL_H__
